+++ changelog.md
@@ -1,6 +1,32 @@
-# version 1.1.0
-- Added command feature to create display block
-- Optimised some CLI internal functionalities
-- Optimised Zlight code to use the latest version of vilhub/router library 
-- Fix service creating issues
-- Fix PHP 8+ backward compatibility issues
+# Version 1.2.0
+
+## Added
+- Database seeding CLI feature
+- Added feature which makes newly added classes via command line to be visible to composer
+
+## Optimised
+- Request Class
+- Error Handling logic
+- pretty_print() function with Kint Library
+
+
+## Fixed
+- Fixed empty error message display for some error types
+- DB certificate check logic issue
+
+# Version 1.1.0
+
+## Added
+- Command feature to create display block
+
+
+## Optimised
+- Some CLI internal functionalities
+- Zlight code to use the latest version of vilhub/router library 
+
+
+## Fixed
+- Service creation issues
+- PHP 8+ backward compatibility issues
+
+